Add files via upload
I have uploaded the codes and some figures used in my paper. My paper may not be very complete, welcome to any advice about my paper. The following is the abstract.

The genius of this paper is to review some important results in filter theory and make several numerical experiments about the application of filter theory in the finance and robotics. In detail, we separately discuss the linear and non-linear filter theory in the beginning, then apply particle filter, Monte-Carlo filter and extended Kalman filter in the filtering problems of pose and volatility estimation. In addition, we use EM (Expectation-Maximum) algorithm to analytically infer the unknown parameters in filtering problems and apply Monte-Carlo simulation to numerically infer theses parameters. Moreover, we introduce the RNNs (recurrent neural networks) and construct RNNs filter to deal with the same numerical experiments of pose and volatility estimation. Finally, we compare the performance of different filters and summarize some possible improvement methods about these filters.